refactor: VariableListContainer 컴포넌트 삭제 > useVariablesList 훅으로 변수 리스트 관리
